Change create_oemof_model to add_to_oemof_model for each component
Every component now implements add_to_oemof_model instead of create_oemof_model. A solph EnergySystem is required.
In the function, the oemof component is created as before, added to the given energy system and returned. In case of more complex components, a tuple may be returned.
Added docstrings and changed documentation to reflect changes.